Removed compile warnings

+++ src/assetHandling/Font.cc
@@ -18,12 +18,12 @@ Font FontLoadFromFile(SDL_Renderer *renderer, const char *path) {
     SDL_FreeSurface(fontSurface);
 
     for (size_t ascii = ASCII_DISPLAY_LOW; ascii <= ASCII_DISPLAY_HIGH; ++ascii) {
-        const int index = int(ascii - ASCII_DISPLAY_LOW);
-        const int col = index % FONT_COLS;
-        const int row = index / FONT_COLS;
+        const size_t index = ascii - ASCII_DISPLAY_LOW;
+        const size_t col = index % FONT_COLS;
+        const size_t row = index / FONT_COLS;
         font.glyphTable[index] = SDL_Rect{
-            .x = col * FONT_CHAR_WIDTH,
-            .y = row * FONT_CHAR_HEIGHT,
+            .x = int(col * FONT_CHAR_WIDTH),
+            .y = int(row * FONT_CHAR_HEIGHT),
             .w = FONT_CHAR_WIDTH,
             .h = FONT_CHAR_HEIGHT,
         };
@@ -33,15 +33,15 @@ Font FontLoadFromFile(SDL_Renderer *renderer, const char *path) {
 
 void RenderChar(SDL_Renderer *renderer, const Font *font, char c, Vector2<int> pos, float scale) {
     const SDL_Rect dst = {
-        .x = (int)floorf(pos.x),
-        .y = (int)floorf(pos.y),
+        .x = (int)floorf(float(pos.x)),
+        .y = (int)floorf(float(pos.y)),
         .w = (int)floorf(FONT_CHAR_WIDTH * scale),
         .h = (int)floorf(FONT_CHAR_HEIGHT * scale),
     };
 
     assert(c >= ASCII_DISPLAY_LOW);
     assert(c <= ASCII_DISPLAY_HIGH);
-    const size_t index = c - ASCII_DISPLAY_LOW;
+    const size_t index = size_t(c - ASCII_DISPLAY_LOW);
     scc(SDL_RenderCopy(renderer, font->spritesheet, &font->glyphTable[index], &dst));
 }
 
@@ -52,7 +52,7 @@ void RenderTextSized(SDL_Renderer *renderer, Font *font, const char *text, size_
     Vector2<int> pen = pos;
     for (size_t i = 0; i < text_size; ++i) {
         RenderChar(renderer, font, text[i], pen, scale);
-        pen.x += FONT_CHAR_WIDTH * scale;
+        pen.x += int(FONT_CHAR_WIDTH * scale);
     }
 }
